refactor(react): extract useMutation hook to shared lib folder
Move useMutation hook and MutationState type from workflows/hooks.ts to lib/use-mutation.ts for reuse by agent hooks and other internal consumers.
