+++ client-sdks/react/src/workflows/types.ts
@@ -1,25 +1,4 @@
-/**
- * Represents the state of a mutation operation.
- * Mimics the essential return type of react-query's useMutation.
- */
-export interface MutationState<TData, TError extends Error, TVariables> {
-  /** Execute the mutation without waiting for the result */
-  mutate: (variables: TVariables) => void;
-  /** Execute the mutation and return a promise with the result */
-  mutateAsync: (variables: TVariables) => Promise<TData>;
-  /** Whether the mutation is currently executing */
-  isPending: boolean;
-  /** Whether the mutation completed successfully */
-  isSuccess: boolean;
-  /** Whether the mutation failed with an error */
-  isError: boolean;
-  /** The error if the mutation failed, null otherwise */
-  error: TError | null;
-  /** The data returned by the mutation if successful */
-  data: TData | undefined;
-  /** Reset the mutation state to initial values */
-  reset: () => void;
-}
+import type { MutationState } from '../lib/use-mutation';
 
 /**
  * Parameters for creating a workflow run.
@@ -51,14 +30,4 @@ export interface StartWorkflowRunParams {
   input: Record<string, unknown>;
   /** Optional request context to pass to the workflow */
   requestContext?: Record<string, unknown>;
-}
-
-/**
- * Return type for the useExecuteWorkflow hook.
- */
-export interface UseExecuteWorkflowReturn {
-  /** Mutation for creating a workflow run */
-  createWorkflowRun: MutationState<CreateWorkflowRunResult, Error, CreateWorkflowRunParams>;
-  /** Mutation for starting a workflow run */
-  startWorkflowRun: MutationState<void, Error, StartWorkflowRunParams>;
 }
